........
  r536480 | rgodfrey | 2007-05-09 11:24:13 +0100 (Wed, 09 May 2007) | 1 line
  
  QPID-482 : Small performance enhancements
........


git-svn-id: https://svn.apache.org/repos/asf/incubator/qpid/trunk@536560 13f79535-47bb-0310-9956-ffa450edef68

+++ qpid/java/common/src/main/java/org/apache/qpid/framing/EncodingUtils.java
@@ -157,7 +157,7 @@ public class EncodingUtils
         }
         else
         {
-            return (short) (1 + s.length());
+            return (1 + s.length());
         }
     }
 
@@ -1023,4 +1023,7 @@ public class EncodingUtils
 
         return l;
     }
+
+    
+
 }
